# docs/swagger.yaml
basePath: /
...
info:
  contact: {}
  description: Simple chat server (Gin + Gorm)
  title: Bubble Chat API
  version: 0.1.0
paths:
...
  /api/guilds/{id}/members:
    get:
      consumes:
      - application/json
      description: 列出服务器成员列表，支持分页和游标分页
      parameters:
      - description: Guild ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 10, max 100)
        in: query
        name: limit
        type: integer
      - description: 'Cursor: return members with id < beforeId'
        in: query
        name: beforeId
        type: integer
      - description: 'Cursor: return members with id > afterId'
        in: query
        name: afterId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{\"data\": [], \"pagination\": {}}'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 非成员
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List guild members
      tags:
      - members
...
  /api/guilds/{id}/members/search:
    get:
      consumes:
      - application/json
      description: 搜索服务器成员（通过名称/昵称）
      parameters:
      - description: Guild ID
        in: path
        name: id
        required: true
        type: integer
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: Limit (default 10, max 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{\"data\": []}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 非成员
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Search guild members
      tags:
      - members
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"